+++ packages/sit-onyx/.storybook/main.ts
@@ -1,4 +1,5 @@
 import type { StorybookConfig } from "@storybook/vue3-vite";
+import { mergeConfig } from "vite";
 
 const config: StorybookConfig = {
   stories: ["./pages/*.mdx", "../src/**/*.stories.ts"],
@@ -13,9 +14,6 @@ const config: StorybookConfig = {
   core: {
     disableTelemetry: true,
   },
-  docs: {
-    autodocs: true,
-  },
   managerHead: (head) => `
     ${head}
     <meta property="og:type" content="website">
@@ -29,6 +27,15 @@ const config: StorybookConfig = {
     <meta property="og:image:height" content="600" />
     <meta property="og:url" content="https://storybook.onyx.schwarz">
   `,
+  // fix "The file does not exist at "..." which is in the optimize deps directory"
+  // see: https://github.com/storybookjs/storybook/issues/28542#issuecomment-2268031095
+  viteFinal: (config) => {
+    return mergeConfig(config, {
+      optimizeDeps: {
+        exclude: ["node_module/.cache/sb-vite"],
+      },
+    } satisfies typeof config);
+  },
 };
 
 export default config;