+++ packages/sit-onyx/src/components/OnyxCheckboxGroup/OnyxCheckboxGroup.stories.ts
@@ -12,6 +12,15 @@ const meta: Meta<typeof OnyxCheckboxGroup> = {
   argTypes: {
     withCheckAll: { control: { type: "boolean" } },
   },
+  decorators: [
+    (story) => ({
+      components: { story },
+      template: `
+      <div style="padding-left: 1rem;">
+        <story />
+      </div>`,
+    }),
+  ],
 };
 
 export default meta;
@@ -28,8 +37,8 @@ const DEMO_OPTIONS = [
     label: "Invalid",
     value: 7,
     customError: {
-      shortMessage: "Example custom error",
-      longMessage: "This text might inform the users what they can do to fix the error.",
+      shortMessage: "Custom error",
+      longMessage: "Further explanation.",
     },
   },
 ] satisfies CheckboxGroupOption[];
@@ -39,7 +48,7 @@ const DEMO_OPTIONS = [
  */
 export const Default = {
   args: {
-    headline: "Checkbox group headline",
+    label: "Checkbox group label",
     modelValue: [2, 6],
     options: DEMO_OPTIONS,
   },