+++ packages/sit-onyx/src/components/OnyxNavBar/modules/OnyxNavButton/NavButtonLayout.vue
@@ -3,16 +3,32 @@
 // to easily switch between mobile and desktop layout
 import arrowSmallLeft from "@sit-onyx/icons/arrow-small-left.svg?raw";
 import { toRef } from "vue";
-import { useManagedState } from "../../../../composables/useManagedState";
+import {
+  MANAGED_SYMBOL,
+  useManagedState,
+  type ManagedProp,
+} from "../../../../composables/useManagedState";
 import { injectI18n } from "../../../../i18n";
 import OnyxButton from "../../../OnyxButton/OnyxButton.vue";
 import OnyxFlyoutMenu from "../OnyxFlyoutMenu/OnyxFlyoutMenu.vue";
 import OnyxNavSeparator from "../OnyxNavSeparator/OnyxNavSeparator.vue";
 import type { OnyxNavButtonProps } from "./types";
 
-const props = withDefaults(defineProps<OnyxNavButtonProps & { isMobile: boolean }>(), {
-  mobileChildrenOpen: false,
-});
+const props = withDefaults(
+  defineProps<
+    OnyxNavButtonProps & {
+      isMobile: boolean;
+      /**
+       * Controls the open state for the mobile children.
+       * Is managed internally if not provided.
+       */
+      mobileChildrenOpen: ManagedProp<boolean>;
+    }
+  >(),
+  {
+    mobileChildrenOpen: MANAGED_SYMBOL,
+  },
+);
 
 const slots = defineSlots<{
   button?(): unknown;
@@ -22,7 +38,7 @@ const slots = defineSlots<{
 const { t } = injectI18n();
 
 const emit = defineEmits<{
-  "update:mobileChildrenOpen": [boolean];
+  "update:mobileChildrenOpen": [isOpen: boolean];
 }>();
 
 const { state: mobileChildrenOpen } = useManagedState(