+++ packages/sit-onyx/src/components/OnyxRadioButton/OnyxRadioButton.stories.ts
@@ -57,4 +57,13 @@ export const Required = {
     name: "radio-required",
     required: true,
   },
+  decorators: [
+    (story) => ({
+      components: { story },
+      template: `
+      <div style="padding: 2rem 0 0 2.5rem;">
+        <story />
+      </div>`,
+    }),
+  ],
 } satisfies Story;