+++ packages/sit-onyx/src/components/OnyxSwitch/OnyxSwitch.stories.ts
@@ -40,6 +40,15 @@ export const Required = {
     ...Default.args,
     required: true,
   },
+  decorators: [
+    (story) => ({
+      components: { story },
+      template: `
+      <div style="padding: 2rem 0 0 1rem;">
+        <story />
+      </div>`,
+    }),
+  ],
 } satisfies Story;
 
 /**
@@ -94,8 +103,17 @@ export const CustomError = {
     ...Default.args,
     hideLabel: true,
     customError: {
-      shortMessage: "Example custom error",
-      longMessage: "This text might inform the users what they can do to fix the error.",
+      shortMessage: "Custom error",
+      longMessage: "Further explanation.",
     },
   },
+  decorators: [
+    (story) => ({
+      components: { story },
+      template: `
+      <div style="padding: 2rem 0 0 4rem;">
+        <story />
+      </div>`,
+    }),
+  ],
 } satisfies Story;